+++ ipc/tests/vmPart_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/multiversx/mx-chain-vm-common-go/mock"
 	"github.com/multiversx/mx-chain-vm-v1_2-go/vmhost"
 	"github.com/multiversx/mx-chain-vm-v1_2-go/config"
-	"github.com/multiversx/mx-chain-vm-v1_2-go/ipc/arwenpart"
+	"github.com/multiversx/mx-chain-vm-v1_2-go/ipc/vmpart"
 	"github.com/multiversx/mx-chain-vm-v1_2-go/ipc/common"
 	"github.com/multiversx/mx-chain-vm-v1_2-go/ipc/marshaling"
 	"github.com/multiversx/mx-chain-vm-v1_2-go/ipc/nodepart"
@@ -81,7 +81,7 @@ func doContractRequest(
 			},
 		}
 
-		part, err := arwenpart.NewArwenPart(
+		part, err := vmpart.NewArwenPart(
 			"testversion",
 			files.inputOfArwen,
 			files.outputOfArwen,
