+++ ipc/vmpart/blockchainGateway_test.go
@@ -1,4 +1,4 @@
-package arwenpart
+package vmpart
 
 import (
 	"fmt"
