+++ ipc/vmpart/messenger.go
@@ -1,4 +1,4 @@
-package arwenpart
+package vmpart
 
 import (
 	"os"
